# lib/api-zod/src/generated/types/riskPrediction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SteelGuard AI simulated industrial safety command center API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskPrediction {
  workerId: string;
  workerName: string;
  riskScore: number;
  riskLevel: string;
  hazardType: string;
  confidence: number;
  recommendation: string;
  notificationMode: string;
  notificationMessage: string;
  notificationsSent: number;
  auditId: string;
}