# lib/api-zod/src/generated/types/riskPredictionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SteelGuard AI simulated industrial safety command center API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskPredictionInput {
  workerId: string;
  workerName: string;
  temperature: number;
  humidity: number;
  gasLevel: number;
  fatigueScore: number;
  ppeCompliance: number;
  workingHours: number;
  hazardDistance: number;
  previousIncidents: number;
}